refactor(SENTZ-5683): Move RefusingHttpRequester beside its only callers
Both its call sites live in CertificateTests.swift, so the type moves
there as a file-private mock.
